Simplify implementation of default op= assignments
[reviewed by diten]

This commit simplifies the implementation of the default op= operators
by removing the explicit casts in the common cases.  This case used
to be required before we supported ops on small ints and good coercion
rules for bools and enums.  Now that those features have all improved,
we can simplify the implementations of these.

The next step in this effort will be to attempt to implement these using
C-level op= for appropriate types to clean up the generated code and
(potentially) supply the C compiler with a cleaner specification.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21446 3a8e244f-b0f2-452b-bcba-4c88e055c3ca