Take 2: Filing bug against supply a directory name to -o
[Sorry for the noise -- I realized I forgot my .future and tried
to abort the previous commit, but clearly failed]

This bug demonstrates a case that we sometimes accidentally run
into in the test system (most recently with intents.chpl and an
intents/ directory), in which if the name given to a -o argument
matches a pre-existing directory name, we get a confusing result.
A cleaner error message would be better.  Should be an easy fix
if anyone needs a 5-minute fix to their name.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21449 3a8e244f-b0f2-452b-bcba-4c88e055c3ca