Fixing a typo in resolveIntents
[reviewed-ish by vass]

Fixing a cut-and-paste typo in resolveIntents that escaped the
original review and didn't cause any actual problems in testing.
The issue is that I was asking something to be declared as
'const' intent to be resolved as though it was 'blank' intent.
I caught this a day or two after the original commit, but took
awhile to getting around to testing and then ran into the code
freeze.  Vass caught it too a day or two ago from now.  I believe
the reason it didn't cause problems is a combination of (a)
blank often meaning const for most types, (b) const checking not
being very good for complex types anyway, and (c) 'const ref'
and 'const in' not being implemented downstream of this pass
correctly yet in all cases.

My apologies for the mistake.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21483 3a8e244f-b0f2-452b-bcba-4c88e055c3ca