In flattenFunctions, handle the 'this' argument as const and so always pass by value to functions being flattened.
[requested and reviewed by Brad]

A simple case where this happens is this:

class C {
  proc outer() {
    proc inner() {
      ... this ...;
    }
    inner();
  }
}

With this patch, 'this'-the-pointer will be passed to the flattened
inner() by value.

This change is half-baked though because (a) there are other cases
that we want to treat the same way and (b) we forego the check
passableByVal() that we apply to other consts.

Also this change used to reduce some comm counts in this test:
  test/distributions/robust/arithmetic/performance/alloc_all.chpl
but at the moment it's hard to tell, given that comm counts are a bit
disturbed in the trunk already.

An addition to a comment is unrelated to the above change, but
still useful.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21484 3a8e244f-b0f2-452b-bcba-4c88e055c3ca