Filing a copy of the bug that was submitted by Hendrik Kassner to chapel-users
The nature of the bug is that we don't properly pass a record by
reference to a dynamically dispatched method due to inserting
too many temps.  See the .future file for a few more details.

A workaround is to use the --conditional-dynamic-dispatch-limit
flag to switch from a vtable to a conditional dispatch.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21485 3a8e244f-b0f2-452b-bcba-4c88e055c3ca