Add .perftimeout capability
This commit adds a .perftimeout that can be used to specify a timeout
for performance runs.  At present, performance runs will use TIMEOUT
by default, if defined, and .perftimeout as the only means of
overriding TIMEOUT.  A better scheme might be to use .timeout if it
exists and .perftimeout doesn't; but the counterargument is that the
current behavior is more symmetric with what we do with
compopts/execopts.

I'm committing this tonight in order to start getting more performance
runs over the weekend that would otherwise timeout (and to avoid
relying on the TIMEOUT/.timeout mechanism because I don't want to grow
the timeout for the non-performance case).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21493 3a8e244f-b0f2-452b-bcba-4c88e055c3ca