Un-futurizing some tests that were futurized in r21480 but are actually passing.
[Committing on behalf of Tom. Reviewed by Vass.]

These futures came from the HL branch where these tests were failing.
Upon branch merge (r21480) these tests started to succeed.
Therefore un-futurizing them. The .no-local.good is not needed either.

Note they have been known to fail with --baseline independently of r21480.

These are related to the nested_var_iterator bug Tom reported.
The bug may still persist, in that when a break is encountered in
the inner loop, we branch to a label that follows the deallocation.
The error could be masked by having no iterator data specifically
associated with the inner loop, but would resurface if, for example,
we break out of a loop due to an exception....

But for now, we are not leaking, so it seems OK to get rid of the
futures.  The onus is on the evil testmaker to come up with a test
that displays this error more directly.  Unfortunately, not all faults
can be evinced or manifested at the Chapel source level....


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21494 3a8e244f-b0f2-452b-bcba-4c88e055c3ca