Removing inappropriate op= overloads
[reviewed by diten, mppf]

In my last 'primitive op=' commit, I wrote up the += functions for
int, uint, real, and imag, and then copy-pasted them for all the other
op= functions without thinking very hard about whether or not they
made sense.  It turns out that a number of them didn't, so this commit
removes the bad ones.  In particular, it removes:

* *=, /=, %= on imag/imag pairs on the argument that they are not
  closed under these operators (e.g., imag * imag => real).

* &=, |=, ^=, >>=, <<= on real/real and imag/imag pairs, on the
  argument that we don't support these bitwise operators on floating
  point values as a binary operator, so shouldn't as an op= call.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21500 3a8e244f-b0f2-452b-bcba-4c88e055c3ca