Quiet fast-on regressions and log futures against source of trouble
In my recent implementation of op= using primitives, I inadvertently
broke three tests that make sure fast-on optimizations fire in certain
cases: distributions/robust/arithmetic/optimizations/optimizeOnClauses*.chpl

The root cause of these failures is that I made all the op= primitives
automatically reject the fast-on optimization rather than doing the
analysis to determine whether or not they might be candidates (if both
arguments were local to the on-clause, for instance).  Somehow my
testing didn't catch the fact that this broke our array accessors, which
rely on += to compute indices.

This commmit fixes the tests but not the root behavior for more
general cases.  What I did was to rewrite the key parts of the array
access code to use a = a + b rather than a += b (and added a TODO
noting that this could be reverted once the root behavior is fixed).
Because assignments are not implemented using 'ref' arguments (they
predate that), they don't have the same problem w.r.t. the fast-on
optimization.

The longer-term fix is to not widen 'ref' arguments unless they need
to be and to make the analysis of op= assignments more like normal
assignments as far as the fast-on optimization goes.  The first future
here -- refArgIsWide -- is designed to see whether the 'ref' argument
to a procedure results in a wide type or not in the generated code.
Once 'ref' arguments are narrow, it should start passing.  The second
future checks to see whether a simple on-clause is optimized by the
fast-on optimization; it should be fixed once the first future is
fixed and op= primitives are analyzed more like PRIM_MOVE is.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21501 3a8e244f-b0f2-452b-bcba-4c88e055c3ca