Add support for developer --ignore-errors-for-pass flag / use to retire future
[reviewed by vass]

This commit adds support for a developer flag called
--ignore-errors-for-pass.  Like the --ignore-errors flag, it
ignores fatal errors, but only until the end of the pass.
At that point, if any errors have occurred, it halts
compilation with an error.  This can be useful, for example,
to write a single test that generates multiple resolution
errors without having to worry about things exploding in
downstream passes or execution time; and beats writing a
separate test for each resolution error.

I used this commit to retire my imaginary op= future, which
is working now that we have op= implemented as a primitive
and have removed cases that use casts and cases that should
be illegal.  It has 4 illegal uses, but without this flag,
only reported the first before exiting.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21503 3a8e244f-b0f2-452b-bcba-4c88e055c3ca