Changing more op= assignments to = temporarily to quiet regressions
This commit is similar to r21501 but for Block-distributed arrays.  As
with the previous commit, it serves to quiet fast-on regression tests
that made sure that array accesses didn't thwart fast-on
optimizations.  The longer-term fix is to stop widening 'ref'
arguments only when needed and treat the op= cases like = cases in the
optimizeOnClauses pass.

If my testing is correct, this should retire all of the remaining
fast-on tests except for the one that makes sure RA gets a fast-on.
RA would need to similarly be rewritten to use = and ^ in order for
that test to pass.  I wanted to discuss this case with others before
making a change to a released file like that.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21514 3a8e244f-b0f2-452b-bcba-4c88e055c3ca