Comment out some code that is unused now but may be valuable later.
[Effectively reviewed by sungeun -- she recommended the same change in
the new util/chplenv/localeModel script yesterday, and that file was
copied from util/chplenv/comm, so now I'm making the same change there.]

Remove the code that looks at the target platform, because we're not
using it.  But do so by commenting it out rather than removing it,
because we may use it someday.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21516 3a8e244f-b0f2-452b-bcba-4c88e055c3ca