Use a single sublocale ID type throughout the runtime interface.
[Reviewed by sungeun.]

Remove the tasking-specific chpl_task_subLoc_t type, in favor of the
c_sublocid_t we already had.  There's no reason to (and several reasons
not to) have two types to refer to sublocale identifiers in the runtime.
In addition to changing the type, I also changed a couple of special
constants that refer to any sublocale and the current sublocale.  They
were defined in chpl-tasks.h and were called chpl_task_anySubLoc and
chpl_task_currSubLoc.  Now they are defined in chpltypes.h and are
called c_sublocid_any and c_sublocid_curr.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21527 3a8e244f-b0f2-452b-bcba-4c88e055c3ca