Add a script that finds all the multilocale tests and directories
This script prints out the names of the multilocale tests and
directories starting in $CHPL_HOME/test (. if $CHPL_HOME is not set).
A multilocale test is defined as any test that is covered by a
NUMLOCALES or .numlocales file.  It does not look in the file, so
tests with 0 or other invalid numLocales are in this list.

Use this with start_test and the --norecurse flag for all your
multilocale testing needs.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21531 3a8e244f-b0f2-452b-bcba-4c88e055c3ca