Add new performance graphing capability
Add new tools to generate a web page of performance graphs generated
by dygraphs.  This will be run side-by-side with the old version
before retiring it to gain some confidence.

The genGraphs script reads .graph files and generates CSV data files
from the original data (which may span multiple raw data files).  It
also generates some JSON for generating the menus and making calls to
the dygraphs library.

In start_test, I followed the existing logic for generating the
graphs.  It's designed to work when using a single perfdir via
CHPL_TEST_PERF_DIR (as in nightly), as well as the local mode where it
just drops the .dat files in ./perfdat/<machine>.  In both cases, the
html is dumped in <perfdir>/html.  The only slightly confusing part is
in the local case when the .graph file is not in the same directory as
where the tests were run (e.g., STREAM.graph and RA.graph).  In such
cases, the html is dropped in the same directory as the .graph file.

The html directory contains:

   CSVfiles/       # directory of CSV data files (generated)
   graphdata.js    # Performance test info (generated)
   index.html      # (static)
   perfgraph.css   # style sheet (static)
   perfgraph.js    # code (static)

Note that since the old way separated generating the graphs (.gifs)
from generate the web page, the web page was only generated if
CHPL_TEST_PERF_DIR was defined.  For this new capability, if you
specify CHPL_TEST_PERF_DIR, start_test will read the graphfiles from
$CHPL_HOME/test/GRAPHFILES.  This allows us to define suites of tests.
The format of GRAPHFILES is as follows:

   # suite: <suite name 1>
   < .graph file name relative to $CHPL_HOME/test>
   < .graph file name relative to $CHPL_HOME/test>
   ...
   < .graph file name relative to $CHPL_HOME/test>
   # suite: <suite name 2>
   < .graph file name relative to $CHPL_HOME/test>
   ...

If you add a new .graph file, it will not show up in the nightly
graphs unless you add it to this file.

FILES
=====

util/test/perf:
util/test/perf/perfgraph.css:
util/test/perf/perfgraph.js:
util/test/perf/index.html:

  Directory of web page guts.  This is the same for every page and is
  copied to the html directory.

util/test/fileReadHelp.py:

  File reading functions.  I pulled this out into a separate Python
  module, because I often copy it from sub_test.  I should pull it out
  of sub_test too.

util/test/genGraphs:

  Main script described above

test/GRAPHFILES

  List of .graph files and suites as described above


TODOs:

- Add a vertical line on every graph for each release branch date

- Ability to have markers on specific graphs

- Ability to have tests in multiple suites

- Better deal with commas in perfkeys in the .graph file (possibly
  change format to one perfkey per line)

- Change the format of the comments at the top of the original .dat
  files (requested by Brad)

- More error handling in genGraphs

- Have start_test check to make sure all .graph files in the test
  directory are in $CHPL_HOME/test/GRAPHFILES

- Add --enddate option for start_test (requested by Brad/Vass)



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21532 3a8e244f-b0f2-452b-bcba-4c88e055c3ca