Make use of CHPL_LOCALE_MODEL_NUM_SUBLOCALES, if it is set.
[Reviewed by sungeun.]

Change all tasking layers to #include "chpl-locale-model.h" so that we
get any definitions set by the locale model.

Change all tasking layers to check for CHPL_LOCALE_MODEL_NUM_SUBLOCALES
and, if it is set, take its value into account when deciding what to
return from chpl_task_numSubLocales().  This is the constant set by a
locale module that wants to force some particular maximum number of
sublocales (typically 0) on tasking layers that are capable of
supporting more.

While here, also put the Chapel-oriented #includes in order in each
tasking layer, except for the initial chplrt.h.  This makes it easier to
find a particular #include file in the set of those included.

In qthreads tasking, also change chpl_task_numSubLocales() to return the
number of shepherds if CHPL_LOCALE_MODEL_NUM_SUBLOCALES is not set or is
too big.  This isn't the ideal value; we actually want the number of
NUMA domains.  But it will do for now.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21537 3a8e244f-b0f2-452b-bcba-4c88e055c3ca