Two bug fixes for extern blocks.
[Reviewed by Brad and Tom]

First bug - problem working with void * (pointed out by Tim).
Second bug - problem with using extern { #define } symbols in same module as
extern block.


The fix for the first bug is just to wrap void* as the opaque type c_ptr -
which makes more sense anyway that trying to redefine void* each time it is
used.

The second bug is a bit more complicated.
I was getting a used-before-defined error with the simple (and new) test
case test/extern/ferguson/externblock/define2.chpl; but putting the
extern block in a module as in test/extern/ferguson/externblock/define.chpl
solved the problem.

Digging deeper, there were two problems. First, externCResolve was adding
the definitions of #define'd symbols to the *end* of the module. Second,
normalize was giving a use-before-define error with the resulting symbol,
even though I believe code generation was adding it correctly.

This might be pointing to a bug elsewhere in the compiler as I observed
curious behavior (described in my request for review posting
to chapel-developers).




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21541 3a8e244f-b0f2-452b-bcba-4c88e055c3ca