Stop writing spaces and newlines in arrays with binary channels
Bug pointed out by Christopher Nelson.
Reviewed by nobody (as I think this is a trivial fix..).

The problem was that the Writer and Reader classes
provided a method, .binary, that was overloaded by
ChannelWriter and ChannelReader. However, parens-less
methods are not subject to dynamic dispatch, so .binary
always returned false (from the default in Writer/Reader).

This patch just switches it to be .binary(), so that
it can be subject to overriding in subclasses, and
in one or two places caches the result in a local
variable so we don't have to do as much dynamic dispatch.
And, of course, it adds tests to make sure the bug
doesn't reappear.

Note -- section 23.5.5 about readThis/writeThis/readWriteThis
in the spec will need to be updated.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21547 3a8e244f-b0f2-452b-bcba-4c88e055c3ca