Improving the 'make clean' step for llvm directory
[OK'd by mppf]

The LLVM build directory used a different 'make clean' rule than the
other third-party directories.  This brings it in line with the
others, partly for the sake of symmetry; partly to avoid the
possibility of errors if you 'make clean' (from the third-party
directory, for example) without ever having built llvm (in which case
the previous approach would error out).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21575 3a8e244f-b0f2-452b-bcba-4c88e055c3ca