Filing reduced version of bug we're seeing in mandelbrot performance test
The mandelbrot-fancy performance test has been failing ever since it
was committed due to a gcc warning that occurs when compiling with
-O3:

        error: assuming signed overflow does not occur when assuming
        that (X + c) < X is always false

The reduced case (mandelbrot-error.chpl) shows that it doesn't take
much to get this warning; the other case (mandelbrot-error-works.chpl)
shows that it doesn't take much different to get a version that
doesn't trigger the warning.  At this point, I'm pretty surprised that
we haven't seen this more or more often.  Did our gcc version get an
upgrade recently (or did we change the flags we use?)

As noted in the .future file, there are gcc flags we could throw to
silence the warning, or to disable the optimization that generates
the warning.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21576 3a8e244f-b0f2-452b-bcba-4c88e055c3ca