Moving some tests around
   qthreads => chap05
   no-local => chap10

Renamed the nightly cron scripts for the above test to reflect what is
being tested and not the machine name, because that seemed to make
more sense as we move things around.

We can lazily update the rest or revert to the old way if this way
seems more cumbersome.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21577 3a8e244f-b0f2-452b-bcba-4c88e055c3ca