Fix bug in intrinsics compareExchange for reals
[post-commit review by Greg]

Fixed a bug in the implementation of compareExchange for gnu
intrinsics for reals.  A __sync_synchronize() call was omitted.  It is
needed for the same reason it is needed in the fetch* cases.  That is,
we are reading non-volatile values.  This should fix the intermittent
failures of the atomic_vars.chpl test.  Ran the tst in a loop for ~30
minutes with no failures (each test takes almost no time..  LOL: not
immeasurable :).

Changed the error message in atomics_vars.chpl to be more specific
about failures and also made the exchange() test a little more
strenuous.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21579 3a8e244f-b0f2-452b-bcba-4c88e055c3ca