Add check to make sure all .graph files are in $CHPL_HOME/test/GRAPHFILES
[reviewed by Lydia]

This patch includes a new script checkGraphfiles (and a small mod to
start_test) that warns when a .graph file does not show up in
$CHPL_HOME/test/GRAPHFILES.  The check is done across the entire test
directory when ever graphs are being generated (similar to the .dat
file name check).  It could be done on a .graph file by .graph file
basis when running in a more "local" mode, but I did not do that
because it would require more csh than I was willing to write at the
moment.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21588 3a8e244f-b0f2-452b-bcba-4c88e055c3ca