Fixing a bug where updates to fields of 'this' are discarded. This used to happen when 'this' is of a record type and the updates are in a task function (begin/on/...) within a constructor for that record. Also filing a bug future for heap-allocating a variable but leaving it uninitialized.
[bug reported and patch reviewed by Sung]

Here is the scenario in question:

  record RR {
    var FIELD...;
    proc RR(...) {         // constructor
      on/begin/... {       // task function
        this.FIELD = ...;  // this update is dropped
      }
    }
  }

The flattenFunctions pass used to pass 'this' by value into
the task function, which, for records, means that any updates to it
would not be visible in the caller's 'this'.

This does not happen for nested functions in a method - because
'this' gets converted to by-reference during resolution in resolveFormals():

          makeRefType(formal->type);
          formal->type = formal->type->refType;

In a constructor, however, 'this' is a local variable, not a formal,
so the above code does not fire.

My fix is to disable r21484 when 'this' is of a record type.
Beware of these considerations:

- This fix makes the behavior of passByRef() (more) consistent
between 'this' and other const enclosing-scope symbols. That is,
the decision is made based on the symbol's type in both cases.

- This fix relies on passableByVal() being 'false' for records.  Since
that can change, this fix is not reliable going forward.  I did not
seek a more principled solution, however, because (a) I am adding a
test to ensure this issue is brought to attention, and (b) Tom's
intention seems to make constructors take 'this' as an argument, in
which case this fix won't be needed in the first place. This change,
however, will remain reasonable because of the first consideration above.


While testing, I encountered a seemingly separate bug, which I am
documenting in this-in-begin-in-method.future.

That bug prevents this-in-taskfns-in-ctors.chpl from working for --local,
so I am adding a skipif and putting this-in-taskfns-in-ctors.chpl
in the multilocale directory to increase the chances of it getting tested.
When the this-in-begin-in-method.* bug is fixed, this-in-taskfns-in-ctors.*
can be moved, say, under test/parallel/taskPar.
this-in-taskfns-in-ctors.skipif can then be removed, and so can
all of this-in-begin-in-method.* because the latter is covered by the former.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21602 3a8e244f-b0f2-452b-bcba-4c88e055c3ca