Adding a bug future where iterator returning an array does so by value vs. by reference depending on whether the return type is specified explicitly.
From .future:

The intention of this future is to require that by-ref-ness do not
depend on whether the iterator's return type is specified in the source
program or inferred by the compiler. In the simplest case, these two
iterators:

  iter iter1() : [D]int { var A: [D]int; yield A; }
  iter iter2()          { var A: [D]int; yield A; }

should either both return 'A' by value or both by reference.

The .good corresponds to the "both return 'A' by value" scenario,
but the other option could be the answer, too.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21604 3a8e244f-b0f2-452b-bcba-4c88e055c3ca