Added a pass to perform loopInvarientCodeMotion. The pass computes the dominators of a function, and uses them to identify all of the natural loops. For each loop, the invariants are computed and are then moved to the loops "pre-header"
The original motivation for adding this pass is because code motion will be required for the initial GPU effort. Currently there is a lot of array/domain meta data present in loops. Beyond the performance penalties this is an issue because openACC, which will be used to target accelerators, can not perform deep copies of structures (as of version 1.0.) Thus the array meta data cannot be copied to the gpu with openacc so it needs to be placed before the loop

I still have to do a bit of cleanup (edit lines down to < 80 characters) but I'm saving that for early next week so it gives me a chance to review the code again after not having looked at it for a bit.


- compiler/include/driver.h
   * Added flag to perform/not perform code motion 
   
- compiler/include/log.h
   * Added my pass 
   
- compiler/include/passes.h
   * Added my pass
   
- compiler/main/driver.cpp
   * Added help information for my pass
   * My pass will run with --fast, will not run with --baseline
   
- compiler/main/passlist.h
   * Added my pass near the end since it has to run after insert wide refs
   
- compiler/optimizations/Makefile.share
   * Added my pass 
   
- compiler/optimizations/loopInvariantCodeMotion.cpp
   * The actual pass to perform the code motion as described above
   
- man/chpl.txt
   * Updated the man page since my pass has a user visible flag 
   
- test/compflags/bradc/help/userhelp.good
   * Updated the help.good since the flag for my pass changed it



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21614 3a8e244f-b0f2-452b-bcba-4c88e055c3ca