Factor insertFormalTemps() in functionResolution.cpp, and add comments related to compiler-added memory management. [reviewed by sungeun]
DETAILS:
    compiler/passes/functionResolution.cpp
Factored addLocalCopiesAndWritebacks() out of insertFormalTemps() to clarify the purpose of the latter, and added many comments describing its behavior.  Replaced the chained if ... else if ... control structure with a switch, to clarify what code is inserted for the various intents.  The conditional involving getRecordWrappedFlags() could be simplified because the cases being rejected are impossible anyway (by visual correlation with the cases that can reach it through the switch). 
The conditional on line 2839 was simplified using the logical identity (!a || (a && b)) == (!a || b).
Added more comments to aid understanding and point out possible simplifications of the code.

    compiler/passes/normalize.cpp
Added comments to interpret the behavior of call_constructor_for_class(). 
Added comments in fix_def_expr() regarding the conditions under which FLAG_INSERT_AUTO_DESTROY is applied.  Under certain conditions noted at the end of the file, clauses may be removed without changing the behavior of the compiler.

    compiler/resolution/wrappers.cpp
Added a comment in buildDefaultWrapper() related to memory management.

    modules/internal/ChapelBase.chpl
The function chpl__autoCopy does not have to be defined for types, so I removed this definition.  The definition for most types is only necessary due to their potential use by instantiate_tuple_autoCopy().  If this function used the same logic as insertFormalTemps() to decide between chpl__initCopy() chpl__autoCopy(), then chpl__autoCopy() would only have to be defined for reference-counted types and tuples containing them.

    modules/internal/DefaultRectangular.chpl
Noted that the DefaultRectangular distribution acts like a singleton, which is why its dsiClone() function does nothing.

    compiler/include/type.h
Added a comment regarding the use fo the "initializer" field.

    compiler/AST/build.cpp
Typo: lbock -> block

[tests: std/all]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21628 3a8e244f-b0f2-452b-bcba-4c88e055c3ca