Microbenchmarks for communication optimization.
Each of these tests has a very similar structure.
They each have a for loop going over an array. The
test either reads or writes to the array elements,
which are a variety of types.

I created these tests to check the functionality
of the LLVM communication aggregation pass, which
I'm about to bring up for review.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21631 3a8e244f-b0f2-452b-bcba-4c88e055c3ca