This is a patch to look for nested synchronization variables in LICM.
It adds two functions. One to find all the nested function symbols and one 
that is a re-factoring of the previous checks for synchronization variables. Other then that, it adds timing for the nested call part and removes some 
test code I left in at the bottom. 

This commit should fix the iterator/angeles regression that currently occurs
in performance testing, though only by coincidence. The real issue with that
is aliases coming from function calls and the patch for that will be coming
shortly. 

As a note, the looking for synchronization variables is very similar to 
what is currently being done in remote value forwarding. It would be a good 
idea to unify the two implementations. 

Other then the above, this patch gets rid of some test code I left at the 
bottom, and adds timing for checking for synchronization variables. 


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21636 3a8e244f-b0f2-452b-bcba-4c88e055c3ca