Refactoring QIO code in order to make it compatible with other filesystems.
Refactored code mainly in qio.h and qio.c, sys.c/h only had one function changed from
static to non-static and renamed (_iov_total_bytes to sys_iov_total_bytes). 

In qio.h several new types, and one new struct type were added and are documented in
the header file. 

The main type that was added was qio_file_funtions_t, which holds all the function
pointers for a configured filesystem as well as the configured FS itself. This is 
populated by a user defined routine that will call into C/runtime from module level 
code (you'll see more in the next commit for HDFS). 

This now leads into the two new fields added to qio_file_t. The two new fields are:
 - qio_file_functions_t* fsfns
 - void* info
Where the fsfns field of the file is set in qio_file_init(_usr). If there is no foreign filsystem,
this field is set to NULL, the standard code path is followed throughout qio.c, and the standard
system functions are called. The info field of the qio_file_t struct holds the
user defined struct that will hold all the information that the library writer might
need in writing their system calls. It is up to the programmer how many fields/what
type of fields they put into this struct (and they are responsible for populating and
packing and unpacking to and from it inside their functions).

New functions that have been added in qio.c for the foreign filesystem work are
always ended in a _usr. Function signatures to the various read and write syscalls have been
modified in order to take in the file struct that they are being called on. 

choose_io_method has been added to in order to check if a foreign filesystem is
present and set the method in a way slightly different from what the normal path is
if so (again if fsfns == NULL it follows the standard path).


TODO's/FIXME's
Future work and TODO's still remain. One thing that Michael had expressed (and which
I haven't gotten around to yet) is changing qio_shortest_path to only take in a
qio_file_funtions_t struct instead of a qio_file_t struct so that getting the path to
a file is not dependent upon opening that file first. I will probably change
this soon, but at least for now it has stayed the way it is. (I'll change it once I
get everything committed up). 

The only thing in this commit that is in any way specifically oriented towards HDFS,
is in qio_file_open_usr, in which if (flags & O_WRONLY) s->seek = NULL; 
This is HDFS specific (since seeking is disallowed in files opened in write mode). So
this should probably get changed at some point in the future, so the opn function
will set this.

A PDF of the documentation for both the foreign filesystem API and the other API's as
well as future work is almost complete and will be put in a site to be determined 
(look at the future HDFS commit to see where this will be).


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21637 3a8e244f-b0f2-452b-bcba-4c88e055c3ca