Simplified parse and build code for tuples. [reviewed by: Lydia and David]
The changes to the parser basically make it obious that the syntax
    '(' <expr> ')' is a parenthesized expression,
    '(' <expr> ',' ')' is a one-tuple,
and parenthesized, comma-separated lists of 2 or more expressions are tuples.  
Since these cases are laid out explicitly in the grammar, the buildParenExpr() function can be simplified as it then handles only the two-or-more-tuple case.

Details:

    chapel.ypp
Inlined <tuple_guts> into <parenthesized_expr>, and re-wrote the action code, so buildParenExpr() only needs to handle the two-or-more-tuple case.  
Renamed it as buildTuple().  
Changed the semantics of tuple_expr_ls to always mean "two or more <tuple_component>s".

    build.cpp
Removed commented out buildZeroTuple().
Simplified buildOneTuple() by removing the PRIM_ACTUALS_LIST call. (see explanation under externCResolve.cpp)
Pushed logic out of buildParenExpr() and into the action code in chapel.ypp.  Renamed it as buildTuple().

    build.h
Renamed buildParenExpr() as buildTuple().

    externCResolve.cpp
Removed an unnecessary PRIM_ACTUALS_LIST call.
(The PRIM_ACTUALS_LIST primitive is used to create a list of actual arguments, but a call to AList::insertAtHead() or AList::insertAtTail() inlines the arguments of PRIM_ACTUALS_LIST into the parent call.  
When creating a one-tuple, the code
  new CallExpr("_build_tuple", new CallExpr(PRIM_ACTUALS_LIST, <args>))
is exactly equivalent to
  new CallExpr("_build_tuple", <args>)
.

[tests run: std-all]



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21638 3a8e244f-b0f2-452b-bcba-4c88e055c3ca