Move RootLocale functionality into ChapelLocale and LocaleModel.
Pushed the RootLocale code into ChapelLocale and LocaleModel depending on if
each locale model needs to define it. This required updates of module uses in
DefaultRectangular and ChapelArray in order to resolve some circularities.

Renamed emptyLocaleSpace and emptyLocales with a leading "chpl_" in
flat/LocaleModel.chpl. Made them local to the virtual functions that use them
in ChapelLocale.chpl.

Also, moved the chpl_localeID_t record into the LocaleModel module since each
locale model needs to define this type for itself. I had to remove the type
aliases used in this record because this type is needed before LocaleModel's
init function is resolved.  I used int(32) directly instead.

[reviewed by Sung]


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21639 3a8e244f-b0f2-452b-bcba-4c88e055c3ca