Minor LLVM backend refactors, add remote memory op aggregation LLVM pass
[Reviewed by Tom]

First, this patch does some minor refactoring to allow some of the LLVM utility
functions to be used from LLVM passes that might not be compiled with
the Chapel compiler (in fact compiler/util/llvm-global-to-wide is
an environment in which llvmAggregateGlobalOps.cpp and llvmGlobalToWide.cpp
are tested with LLVM unit tests with basically no Chapel dependencies).

Then, it adds and enables llvmAggregageGlobalOps.cpp which is an LLVM
pass to combine loads or stores on global memory when the pass can prove
that they are contiguous. For example, the pass might transform

  %p0 = getelementptr inbounds i64 addrspace(100)* %base, i32 3
  %v1 = load i64 addrspace(100)* %p0, align 8
  %p1 = getelementptr inbounds i64 addrspace(100)* %base, i32 1
  %v2 = load i64 addrspace(100)* %p1, align 8

(address space 100 means Global Pointer)

into something that does the Global operation with a memcpy to
a stack-local variable and then does the loads out of that.
Since loads/stores/memcpys translate into puts/gets, the single
memcpy for mulitple loads or stores translates into a reduction
in the number of puts or gets. Works for loads and stores, works
for loads even with gaps.

The algorithm (and code) for llvmAggregageGlobalOps draws heavily from
the memset optimization algorithm in LLVM's MemCpyOpt pass.

With this patch, compiling with --llvm --llvm-wide-opt --fast
has same or lower comms counts for all of the communication
microbenchmarks in test/performance/ferguson. This aggregation
pass resolves a performance issue with --llvm-wide-opt where
LLVM memory optimizations would turn small memcpys on global
memory (ie how llvm-wide-opt represents puts and gets) into
many small loads and stores (ie more puts and gets).

For example, with test/performance/ferguson/remote-tuple-write-small-1.chpl
Normal:
 (get = 0, get_nb = 0, get_nb_test = 0, get_nb_wait = 0, put = 0, fork = 1, fork_fast = 0, fork_nb = 0) (get = 200012, get_nb = 0, get_nb_test = 0, get_nb_wait = 0, put = 300000, fork = 0, fork_fast = 0, fork_nb = 0)

LLVM Wide Opt + Aggregation:
 (get = 0, get_nb = 0, get_nb_test = 0, get_nb_wait = 0, put = 0, fork = 1, fork_fast = 0, fork_nb = 0) (get = 100004, get_nb = 0, get_nb_test = 0, get_nb_wait = 0, put = 100000, fork = 0, fork_fast = 0, fork_nb = 0)

Future work - fix LLVM TBAA for Chapel, make the aggregation pass
appropriately use TBAA, improve the code generation of wide
pointers so that LLVM passes can easily know when a wide pointer
is really actually referring to a stack local variable - or improve
the Chapel code generator to not make wide pointers for local
operations on stack local variables... (I'm seeing this in
particular for test/performance/remote-record-read.chpl). 



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21640 3a8e244f-b0f2-452b-bcba-4c88e055c3ca